+++ .github/workflows/security.yml
@@ -1,0 +1,15 @@
+name: Security Check
+
+on: [push, pull_request]
+
+jobs:
+    security:
+        if: ${{ secrets.SNYK_TOKEN != '' }}
+        runs-on: ubuntu-latest
+        continue-on-error: true
+        steps:
+            - uses: actions/checkout@v4
+            - name: Run Snyk to check for vulnerabilities
+              uses: snyk/actions/node@master
+              env:
+                  SNYK_TOKEN: ${{ secrets.SNYK_TOKEN }}
