+++ src/starbunk/bots/reply-bots/blueBot.ts
@@ -1,6 +1,6 @@
 import userID from '@/discord/userID';
 import { OpenAIClient } from '@/openai/openaiClient';
-import { Logger } from '@/services/Logger';
+import { Logger } from '@/services/logger';
 import ReplyBot from '@/starbunk/bots/replyBot';
 import { WebhookService } from '@/webhooks/webhookService';
 import { Message, TextChannel } from 'discord.js';
