+++ src/starbunk/bots/reply-bots/sheeshBot.ts
@@ -1,4 +1,4 @@
-import { Logger } from '@/services/Logger';
+import { Logger } from '@/services/logger';
 import ReplyBot from '@/starbunk/bots/replyBot';
 import { WebhookService } from '@/webhooks/webhookService';
 import { Message, TextChannel } from 'discord.js';
