+++ src/starbunk/commands/play.ts
@@ -1,4 +1,4 @@
-import { Logger } from '@/services/Logger';
+import { Logger } from '@/services/logger';
 import StarbunkClient, { getStarbunkClient } from '@/starbunk/starbunkClient';
 import { AudioPlayerStatus, joinVoiceChannel } from '@discordjs/voice';
 import { ChatInputCommandInteraction, GuildMember, SlashCommandBuilder } from 'discord.js';
